Completed the desktop required and optional application installer, some application on fedora will be installed using snap instead

+++ install/desktop/optional/app-steam.sh
@@ -1,6 +1,11 @@
-# Play games from https://store.steampowered.com/
-cd /tmp
-wget https://cdn.akamai.steamstatic.com/client/installer/steam.deb
-sudo apt install -y ./steam.deb
-rm steam.deb
-cd -
+if [ "$OMAKUB_OS" = "ubuntu" ]; then
+  # Play games from https://store.steampowered.com/
+  cd /tmp
+  wget https://cdn.akamai.steamstatic.com/client/installer/steam.deb
+  sudo apt install -y ./steam.deb
+  rm steam.deb
+  cd -
+elif [ "$OMAKUB_OS" = "fedora" ]; then
+  # Play games from https://store.steampowered.com/
+  sudo snap install steam
+fi
